+++ decidim-meetings/app/assets/config/admin/decidim_meetings_manifest.js
@@ -1,1 +1,1 @@
-//= link decidim/meetings/admin/inscriptions_form.js
+//= link decidim/meetings/admin/registrations_form.js