+++ decidim-meetings/app/models/decidim/meetings/meeting.rb
@@ -12,7 +12,7 @@ module Decidim
       include Decidim::HasScope
       include Decidim::HasCategory
 
-      has_many :inscriptions, class_name: "Decidim::Meetings::Inscription", foreign_key: "decidim_meeting_id"
+      has_many :registrations, class_name: "Decidim::Meetings::Registration", foreign_key: "decidim_meeting_id"
 
       feature_manifest_name "meetings"
 
@@ -26,15 +26,15 @@ module Decidim
 
       def has_available_slots?
         return true if available_slots.zero?
-        available_slots > inscriptions.count
+        available_slots > registrations.count
       end
 
       def remaining_slots
-        available_slots - inscriptions.count
+        available_slots - registrations.count
       end
 
-      def has_inscription_for?(user)
-        inscriptions.where(user: user).any?
+      def has_registration_for?(user)
+        registrations.where(user: user).any?
       end
     end
   end