+++ decidim-meetings/app/models/decidim/meetings/registration.rb
@@ -2,8 +2,8 @@
 
 module Decidim
   module Meetings
-    # The data store for a Inscription in the Decidim::Meetings component.
-    class Inscription < Meetings::ApplicationRecord
+    # The data store for a Registration in the Decidim::Meetings component.
+    class Registration < Meetings::ApplicationRecord
       belongs_to :meeting, foreign_key: "decidim_meeting_id", class_name: "Decidim::Meetings::Meeting"
       belongs_to :user, foreign_key: "decidim_user_id", class_name: "Decidim::User"
     end