+++ decidim-meetings/spec/commands/join_meeting_spec.rb
@@ -3,9 +3,9 @@
 require "spec_helper"
 
 describe Decidim::Meetings::JoinMeeting do
-  let(:inscriptions_enabled) { true }
+  let(:registrations_enabled) { true }
   let(:available_slots) { 10 }
-  let(:meeting) { create :meeting, inscriptions_enabled: inscriptions_enabled, available_slots: available_slots }
+  let(:meeting) { create :meeting, registrations_enabled: registrations_enabled, available_slots: available_slots }
   let(:user) { create :user, :confirmed, organization: meeting.organization }
   subject { described_class.new(meeting, user) }
 
@@ -14,16 +14,16 @@ describe Decidim::Meetings::JoinMeeting do
       expect { subject.call }.to broadcast(:ok)
     end
 
-    it "creates an inscription for the meeting and the user" do
-      expect { subject.call }.to change { Decidim::Meetings::Inscription.count }.by(1)
-      last_inscription = Decidim::Meetings::Inscription.last
-      expect(last_inscription.user).to eq(user)
-      expect(last_inscription.meeting).to eq(meeting)
+    it "creates an registration for the meeting and the user" do
+      expect { subject.call }.to change { Decidim::Meetings::Registration.count }.by(1)
+      last_registration = Decidim::Meetings::Registration.last
+      expect(last_registration.user).to eq(user)
+      expect(last_registration.meeting).to eq(meeting)
     end
   end
 
-  context "when the meeting has not inscriptions enabled" do
-    let(:inscriptions_enabled) { false }
+  context "when the meeting has not registrations enabled" do
+    let(:registrations_enabled) { false }
 
     it "broadcasts invalid" do
       expect { subject.call }.to broadcast(:invalid)
@@ -34,7 +34,7 @@ describe Decidim::Meetings::JoinMeeting do
     let(:available_slots) { 1 }
 
     before do
-      create(:inscription, meeting: meeting)
+      create(:registration, meeting: meeting)
     end
 
     it "broadcasts invalid" do