+++ decidim-meetings/spec/commands/leave_meeting_spec.rb
@@ -3,15 +3,15 @@
 require "spec_helper"
 
 describe Decidim::Meetings::LeaveMeeting do
-  let(:inscriptions_enabled) { true }
+  let(:registrations_enabled) { true }
   let(:available_slots) { 10 }
-  let(:meeting) { create :meeting, inscriptions_enabled: inscriptions_enabled, available_slots: available_slots }
+  let(:meeting) { create :meeting, registrations_enabled: registrations_enabled, available_slots: available_slots }
   let(:user) { create :user, :confirmed, organization: meeting.organization }
   let(:user_leaving_meeting) { user }
   subject { described_class.new(meeting, user_leaving_meeting) }
 
   before do
-    create(:inscription, meeting: meeting, user: user)
+    create(:registration, meeting: meeting, user: user)
   end
 
   context "when everything is ok" do
@@ -19,8 +19,8 @@ describe Decidim::Meetings::LeaveMeeting do
       expect { subject.call }.to broadcast(:ok)
     end
 
-    it "destroys the inscription for the meeting and the user" do
-      expect { subject.call }.to change { Decidim::Meetings::Inscription.count }.by(-1)
+    it "destroys the registration for the meeting and the user" do
+      expect { subject.call }.to change { Decidim::Meetings::Registration.count }.by(-1)
     end
   end