+++ decidim-meetings/spec/commands/update_registrations_spec.rb
@@ -2,12 +2,12 @@
 
 require "spec_helper"
 
-describe Decidim::Meetings::Admin::UpdateInscriptions do
+describe Decidim::Meetings::Admin::UpdateRegistrations do
   let(:meeting) { create(:meeting) }
   let(:invalid) { false }
-  let(:inscriptions_enabled) { true }
+  let(:registrations_enabled) { true }
   let(:available_slots) { 10 }
-  let(:inscription_terms) do
+  let(:registration_terms) do
     {
       en: "A legal text",
       es: "Un texto legal",
@@ -17,9 +17,9 @@ describe Decidim::Meetings::Admin::UpdateInscriptions do
   let(:form) do
     double(
       invalid?: invalid,
-      inscriptions_enabled: inscriptions_enabled,
+      registrations_enabled: registrations_enabled,
       available_slots: available_slots,
-      inscription_terms: inscription_terms
+      registration_terms: registration_terms
     )
   end
 
@@ -40,9 +40,9 @@ describe Decidim::Meetings::Admin::UpdateInscriptions do
 
     it "updates the meeting" do
       subject.call
-      expect(meeting.inscriptions_enabled).to eq(inscriptions_enabled)
+      expect(meeting.registrations_enabled).to eq(registrations_enabled)
       expect(meeting.available_slots).to eq(available_slots)
-      expect(translated(meeting.inscription_terms)).to eq "A legal text"
+      expect(translated(meeting.registration_terms)).to eq "A legal text"
     end
   end
 end